% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{app_ui}
\alias{app_ui}
\title{Shiny App UI

app_ui <- function() {
shiny::fluidPage(
# Application title
shiny::titlePanel("Statistical Test App"),}
\usage{
app_ui()
}
\description{
\if{html}{\out{<div class="sourceCode">}}\preformatted{# Use the shinyjs library
shinyjs::useShinyjs(),
}\if{html}{\out{</div>}}
}
\details{

}
