% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{choose_statistical_test}
\alias{choose_statistical_test}
\title{Choose Statistical Test}
\usage{
choose_statistical_test(independent_var, dependent_var)
}
\arguments{
\item{independent_var}{The independent variable.}

\item{dependent_var}{The dependent variable.}
}
\value{
The recommended statistical test.
}
\description{
This function chooses the appropriate statistical test based on the provided table.
}
