% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{create_dependent_variable_histogram}
\alias{create_dependent_variable_histogram}
\title{Create a Histogram Plot for the Dependent Variable}
\usage{
create_dependent_variable_histogram(dependent_var)
}
\arguments{
\item{dependent_var}{A numeric vector representing the dependent variable.}
}
\value{
No return value, the function creates a histogram plot.
}
\description{
This function creates a histogram plot for a numeric dependent variable.
}
