% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{determine_dependent_variable}
\alias{determine_dependent_variable}
\title{Determine the Type of Dependent Variable}
\usage{
determine_dependent_variable(dependent_var)
}
\arguments{
\item{dependent_var}{A vector representing the dependent variable.}
}
\value{
A character string indicating the type of the dependent variable.
}
\description{
This function determines the type of the dependent variable based on its data type and distribution.
}
