% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{determine_independent_variable}
\alias{determine_independent_variable}
\title{Determine the Type of Independent Variable}
\usage{
determine_independent_variable(independent_var)
}
\arguments{
\item{independent_var}{A vector representing the independent variable.}
}
\value{
A character string indicating the type of the independent variable.
}
\description{
This function determines the type of the independent variable based on its data type and the number of unique values.
}
