% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_vvdoctor.R
\name{diagnose}
\alias{diagnose}
\title{Diagnose Data with Statistical Tests}
\usage{
diagnose()
}
\value{
NULL The function does not return a value, but launches the Shiny app.
}
\description{
This function launches a Shiny application that allows users to upload data
and perform various statistical tests on the dataset. Users can select the
dependent and independent variables, choose the appropriate statistical test,
and view the results.
}
\examples{
if (interactive()) {
  diagnose()
}

}
