% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperfunctions.R
\name{display_data_table}
\alias{display_data_table}
\title{Display Data Table}
\usage{
display_data_table(data)
}
\arguments{
\item{data}{The data frame to display.}
}
\description{
Helper function to display a data table in a Shiny app.
}
