% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{get_variable_class_info}
\alias{get_variable_class_info}
\title{Get Variable Class Information}
\usage{
get_variable_class_info(variable)
}
\arguments{
\item{variable}{The selected variable data.}
}
\value{
A character vector containing the variable class and additional information.
}
\description{
This function determines the class of the selected variable and additional information.
}
