% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperfunctions.R
\name{handle_file_upload}
\alias{handle_file_upload}
\title{Handle File Upload and Update Dropdowns}
\usage{
handle_file_upload(input, output, session)
}
\arguments{
\item{input}{The input object from the Shiny app.}

\item{output}{The output object from the Shiny app.}

\item{session}{The session object from the Shiny app.}
}
\value{
A reactive value containing the uploaded data.
}
\description{
Helper function for handling file upload and updating dropdowns in a Shiny app.
}
