% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{perform_statistical_test}
\alias{perform_statistical_test}
\title{Perform Statistical Test}
\usage{
perform_statistical_test(data, input)
}
\arguments{
\item{data}{A data frame containing the variables.}

\item{input}{A list containing the user's input, including the selected statistical test, dependent variable, independent variable, identifier variable, and input mean.}
}
\value{
The result of the statistical test.
}
\description{
This function performs the appropriate statistical test based on the user's input.
}
