% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_vvdoctor.R
\name{run_vvdoctor}
\alias{run_vvdoctor}
\title{Run the Statistical Test App}
\usage{
run_vvdoctor()
}
\description{
This function starts the Shiny app, which allows users to upload data and perform
statistical tests on the dataset.
}
\examples{
\dontrun{
library(vvdoctor)
run_statistical_test_app()
}
}
